requirements.txt
beautifulsoup4
blockdiag
blosc
bqplot
chardet
cloudpickle
coverage
cytoolz
dbfread
duecredit
ete3
folium
forestci
gensim
geopy
ggplot
jsonpickle
jyquickhelper
keras
keyring
lifelines
lightgbm
lightning-python
manydataapi
mlstatpy>=0.1.335
mpld3
mutagen
opencv-python
pandas_streaming
partd
plotly
pocket
psutil
pydub
pyenbc
pyensae>=1.2.731
pygal
pygal_maps_fr
pygame
pylzma
pymmails
pympler
pymyinstall>=1.1.1415
pyqt5
pyquickhelper>=1.8
pyrsslocal
python-jenkins
qgrid
semantic_version
seasonal
shapely
tables
toolz
sphinxjp.themes.basicstrap
sphinxjp.themes.sphinxjp
tensorflow>=1.5
tkinterquickhelper
treeinterpreter
tutormagic
wheel
xlrd
xlsxwriter
xlwt
